Simplify bridge type generation
