+++ package.json
@@ -39,7 +39,7 @@
     "build:linux": "pnpm run build && electron-builder --linux",
     "bootstrap": "node scripts/build-bootstrap-python.mjs",
     "bootstrap:fetch": "node scripts/fetch-bootstrap-python.mjs",
-    "bridge-types:gen": "node scripts/generate-comfy-desktop-bridge-types.mjs",
+    "bridge-types:gen": "pnpm exec tsc -p tsconfig.comfy-desktop-bridge-types.json",
     "font:fetch": "node scripts/fetch-font.mjs",
     "sandbox:gen": "node scripts/generate-sandbox.mjs",
     "datadog:sourcemaps:upload": "node ./scripts/upload-datadog-sourcemaps.mjs",
