+++ tsconfig.comfy-desktop-bridge-types.json
@@ -6,7 +6,7 @@
     "module": "ESNext",
     "moduleResolution": "bundler",
     "noUncheckedIndexedAccess": true,
-    "outDir": "packages/comfyui-desktop-bridge-types/.generated",
+    "outDir": "packages/comfyui-desktop-bridge-types",
     "rootDir": "src/types",
     "skipLibCheck": true,
     "strict": true,
